ref: moved security classes from services to security package
